Создал начальные файлы для работы с каталогом
