+++ FakeNFT/Scenes /Catalog/CatalogModel/CatalogMainScreenModel.swift
@@ -1,0 +1,13 @@
+//
+//  CatalogMainScreenModel.swift
+//  FakeNFT
+//
+//  Created by Олег Серебрянский on 8/13/24.
+//
+
+import Foundation
+import Combine
+
+final class CatalogMainScreenModel {
+    
+}
