+++ FakeNFT/Scenes /Catalog/CustomCells/CustomCellTableView.swift
@@ -1,0 +1,8 @@
+//
+//  CustomCellTableView.swift
+//  FakeNFT
+//
+//  Created by Олег Серебрянский on 8/13/24.
+//
+
+import Foundation
